sqlModel is renamed and userRepository.go deleted
